+++ src/ClientMessageProcessor.h
@@ -24,7 +24,7 @@
 #include "DXPMessageData.h"
 
 bool dx_subscribe_symbols_to_events (dxf_connection_t connection, dx_order_source_array_ptr_t order_source,
-                                     dxf_const_string_t* symbols, int symbol_count, int event_types, bool unsubscribe,
+                                     dxf_const_string_t* symbols, size_t symbol_count, int event_types, bool unsubscribe,
                                      bool task_mode, dxf_uint_t subscr_flags, dxf_long_t time);
 
 bool dx_send_record_description (dxf_connection_t connection, bool task_mode);
