+++ src/DXMemory.h
@@ -32,12 +32,12 @@
  */
 /* -------------------------------------------------------------------------- */
 
-void* dx_malloc (int size);
-void* dx_calloc (int num, int size);
+void* dx_malloc (size_t size);
+void* dx_calloc (size_t num, size_t size);
 void  dx_free (void* buf);
-void* dx_memcpy (void* destination, const void* source, int size);
-void* dx_memmove (void* destination, const void* source, int size);
-void* dx_memset (void* destination, int c, int size);
+void* dx_memcpy (void* destination, const void* source, size_t size);
+void* dx_memmove (void* destination, const void* source, size_t size);
+void* dx_memset (void* destination, int c, size_t size);
 
 /* -------------------------------------------------------------------------- */
 /*
@@ -48,7 +48,7 @@ void* dx_memset (void* destination, int c, int size);
  */
 /* -------------------------------------------------------------------------- */
 
-void* dx_calloc_no_ehm (int num, int size);
+void* dx_calloc_no_ehm (size_t num, size_t size);
 void dx_free_no_ehm (void* buf);
 
 #endif /* DX_MEMORY_H_INCLUDED */
